src/bot/pyproject.toml
[tool.poetry]
name = "glue"
version = "0.1.0"
description = ""
authors = ["Moritz <fullermoritz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
"discord.py" = {git = "git@github.com:Rapptz/discord.py.git"}
pymongo = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
